research/ocr/reference/temi_experiments/detectron_requirements.txt
absl-py==2.0.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
black==21.4b2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
detectron2==0.5
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
hydra-core==1.3.2
idna==3.6
iopath==0.1.9
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
layoutparser==0.3.4
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.12.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.1.0
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydot==1.4.2
Pygments==2.17.2
pyparsing==3.1.1
pypdfium2==4.25.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
toml==0.10.2
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
yacs==0.1.8
